Basic character controller with 3 different views.
The player prefab includes movement, collision, and jumping. A first-person and 3rd-person camera are part of the prefab and already assigned. The Rotation-Lock Camera will provide a top-down view centered on the player.

The top-down camera will have to be assigned to the player prefab as an available camera, and must not be a child of the player prefab.

Rotation was wonky, so that will be addressed later.